refactor: to human date function
